+++ ink_metadata/mod.ts
@@ -1,1 +1,3 @@
+export * from "./known.ts"
 export * from "./Metadata.ts"
+export * from "./normalizeTys.ts"
